+++ changelog.md
@@ -1,1 +1,15 @@
+- **Undetected WoodType**
+    - Added Sniffed-Out, vessel
+    - Added Dungeon's Delight, wormwood
+
+---
+
+## v2.16.13
+
+- Restored the Undetected LeavesType from Mystic's Biomes - [#390](https://github.com/MehVahdJukaar/Moonlight/issues/390)
+- Improved one code in LeavesType's detection system
+- Removed WoodType, `culturaldelights:avocado` because the child of planks is `jungle_planks`
+
+## v2.16.12
+
 - Added Cultural's Delights' undetected WoodType and 1 LeavesType's missing Associated WoodType - [#1050](https://github.com/MehVahdJukaar/WoodGood/issues/1050)